refactor: Rename "swiperGlobalFunctionality"
