+++ js/index.js
@@ -3,5 +3,5 @@ import "./modules/popupRegistration.js";
 import "./modules/swiperOurHorsesFunctionality.js";
 import "./modules/swiperCommentsFunctionality.js";
 import "./modules/sliderOurVisitors.js";
-import "./modules/inputValidation.js";;
+import "./modules/inputValidation.js";
 import "./modules/mapFunctionality.js";
