+++ js/modules/swiperOurHorsesFunctionality.js
@@ -1,4 +1,4 @@
-import { navigationButtonAppearance, setSlidesState } from "./swiperGlobalFunctionality.js";
+import { navigationButtonAppearance, setSlidesState } from "./swiperHelpers.js";
 import Swiper from "./swiper-bundle.esm.browser.min.js";
 
 const swiper = new Swiper(".swiper-container-1", {
